Rewrite udev/systemd integration
In discussions with upstream, we think that since there's a many-to-one
mapping of device nodes to systemd service units, it makes sense to use
systemd/udev ENV{SYSTEMD_WANTS}+= to tie the appearance of a device node
to automatically start the corresponding service unit. Then the service
units have no [Install] section since they are started automatically
when device nodes appear. This should work well on systems where not all
device nodes (if any) are available.

There's some discussion of this at
https://github.com/qualcomm/fastrpc/issues/245

Signed-off-by: Robie Basak <robie.basak@oss.qualcomm.com>