Drop fastrpc-tests as source is not available
Debian requires source to be available, but it currently isn't in the
test suite. To make progress, drop the test suite from the build for
now for Debian. This can be reverted when the issue is resolved upstream
and we can build the entire test suite directly from source.

This is tracked upstream at
https://github.com/qualcomm/fastrpc/issues/235

Signed-off-by: Robie Basak <robie.basak@oss.qualcomm.com>